Feat: starting order controller

+++ src/ChoppSoft.Api/ViewModels/ProductViewModel.cs
@@ -9,7 +9,7 @@ namespace ChoppSoft.Api.ViewModels
         public string Identifier { get; set; }
         public string Description { get; set; }
         public string Brand { get; set; }
-        public double Capacity { get; set; }
+        public decimal Capacity { get; set; }
         public decimal Price { get; set; }
         public Guid? WarehouseId { get; set; }
         public string WarehouseDescription { get; set; }
