+++ src/ChoppSoft.Api/ViewModels/ResourceViewModel.cs
@@ -6,7 +6,7 @@
         public string Description { get; set; }
         public string Model { get; set; }
         public string LicensePlate { get; set; }
-        public double Capacity { get; set; }
+        public decimal Capacity { get; set; }
         public bool IsOwned { get; set; }
     }
 }
